Update: changed type of field "date" in models "Article" and "Comment"

+++ blog/articles/models.py
@@ -1,10 +1,12 @@
 from django.db import models
 
+
 class Article(models.Model):
     title = models.CharField('Заголовок', max_length=50)
     definition = models.CharField('Описание', max_length=140)
     text = models.TextField('Текст')
-    date = models.IntegerField('Дата публикации')
+    date = models.DecimalField(
+        'Дата публикации', max_digits=15, decimal_places=0)
 
     def __str__(self) -> str:
         return self.title
